src/requirements.txt
flask>=2.2.3
flask-markdown>=0.3
markdown>=3.4.1
beautifulsoup4>=4.11.1
python-frontmatter>=1.1.0
requests>=2.32.3
pillow>=11.0.0
